Simplify `GrainPP.Resolver`
griffe (the docs parser used in jet-docs) trips up on the loop of
`GrainPP` -> `GrainPP.Resolver` -> `GrainPP` -> `GrainPP.Resolver` -> etc.
Because of this the whole `vsdeband.noise` module had to be excluded.
While this could be fixed in griffe, it also looked like this could be
both fixed and simplified by just quoting the type (like `'GrainPP'`).
This way there's no loop and no need for `_gpp`.

Also don't need the the explicit `type` in `GrainPostProcessT` (8550361)
anymore.

Verified that the computed type is the same before and after:

    (type) Resolver = (VideoNode) -> GrainPP